reorgnize tests into their own directoy, split out the files a bit
