+++ python/dqmtools/dqmtests/__init__.py
@@ -1,0 +1,3 @@
+from .common_tests import *
+from .wibeth_tests import *
+from .daphne_tests import *
