+++ python/dqmtools/dqmtests/wibeth_tests.py
@@ -4,20 +4,6 @@ from rawdatautils.unpack.dataclasses import *
 import numpy as np
 import operator
 
-class CheckAllExpectedFragmentsTest(DQMTest):
-
-    def __init__(self):
-        super().__init__()
-        self.name = 'CheckAllExpectedFragmentsTest'
-
-    def run_test(self,df_dict):
-        df_tmp = (df_dict["trh"]["n_fragments"]!=df_dict["trh"]["n_requested_components"])
-        n_not_matched = df_tmp.sum()
-        if n_not_matched==0:
-            return DQMTestResult(DQMResultEnum.OK,f'OK')
-        else:
-            return DQMTestResult(DQMResultEnum.BAD,f'{n_not_matched} / {len(df_tmp)} records missing fragments.')
-
 class CheckTimestampDiffs_WIBEth(DQMTest):
     
     def __init__(self,det_name):
@@ -169,53 +155,6 @@ class CheckWIBEth_FEMB_Sync(CheckWIBEth_Header_Value):
         super().__init__(det_name,"femb_sync",3)
         self.name = f'CheckWIBEth_FEMB_Sync_{det_name}'
 
-        
-class CheckTimestampsAligned(DQMTest):
-
-    def __init__(self,det_id,verbose=True):
-        super().__init__()
-        self.det_id = det_id
-        self.name = f'CheckTimestampsAligned_{det_id}'
-        self.verbose = verbose
-        
-    def any_different(arr):
-        return (arr.values!=arr.values[0]).sum()
-
-    def unique(arr):
-        return np.unique(arr.values,return_counts=True)
-    
-    def run_test(self,df_dict):
-        df_tmp = df_dict["daqh"].loc[df_dict["daqh"]["det_id"]==self.det_id]
-        
-        if len(df_tmp)==0:
-            return DQMTestResult(DQMResultEnum.WARNING,f'WARNING: No components found with detector id {self.det_id}.')
-        
-        df_tmp_gb = df_tmp.groupby(by=["run","trigger","sequence"])["timestamp_first_dts"].agg(CheckTimestampsAligned.unique)
-        df_tmp_gb_n = df_tmp_gb.apply(lambda x: len(x[1]))
-        n_different = (df_tmp_gb_n!=1).sum()
-
-        if n_different==0:
-            return DQMTestResult(DQMResultEnum.OK,f'OK')
-        else:
-            if self.verbose:
-                df_tmp_fr = df_dict["frh"].loc[df_dict["frh"]["det_id"]==self.det_id][["trigger_timestamp_dts","window_begin_dts","window_end_dts"]]
-                
-                df_tmp_gb_mode = df_tmp_gb.apply(lambda x: x[0][np.argmax(x[1])])
-                df_tmp = df_tmp.join(df_tmp_gb_mode,rsuffix='_majority')
-                df_tmp = df_tmp.join(df_tmp_fr)
-                df_tmp = df_tmp.loc[(df_tmp["timestamp_first_dts"]!=df_tmp["timestamp_first_dts_majority"])]
-
-                df_tmp["timestamp_diff"] = df_tmp["timestamp_first_dts"]-df_tmp["timestamp_first_dts_majority"]
-                n_different = len(np.unique(df_tmp.reset_index()["src_id"].apply(lambda x: int(x))))
-                print(f"\nFRAGMENTS FAILING TIMESTAMP ALIGNMENT for Detector ID {self.det_id}")
-                print(tabulate(df_tmp.reset_index()[["trigger","sequence","crate_id","slot_id","stream_id","timestamp_first_dts","timestamp_first_dts_majority","timestamp_diff",
-                                                     "window_begin_dts","window_end_dts"]],
-                                headers=["Record","Seq.","Crate","Slot","Stream","Timestamp (first)","Majority timestamp","Difference","Window begin","Window end"],
-                                showindex=False,tablefmt='pretty'))
-
-            return DQMTestResult(DQMResultEnum.BAD,
-                                 f'{n_different} sources have some timestamp misalignment for det_id {self.det_id}.')
-
 class CheckNFrames_WIBEth(DQMTest):
 
     def __init__(self):
@@ -235,7 +174,7 @@ class CheckNFrames_WIBEth(DQMTest):
             return DQMTestResult(DQMResultEnum.BAD,
                                  f'{n_frames_wrong} / {len(df_tmp)} WIBEth fragments have the wrong number of frames.')
 
-class CheckRequestTimes_WibEth(DQMTest):
+class CheckRequestTimes_WIBEth(DQMTest):
 
     def __init__(self,det_name,verbose=True):
         super().__init__()
@@ -371,117 +310,4 @@ class CheckPedestal_WIBEth(DQMTest):
                                showindex=False,tablefmt='pretty',floatfmt=".2f"))
             return DQMTestResult(DQMResultEnum.BAD,
                                  f'{n_bad} channels have pedestal outside of range.')
-
-class CheckEmptyFragments_DAPHNE(DQMTest):
-
-    def __init__(self):
-        super().__init__()
-        self.name = "CheckEmptyFragments_DAPHNE"
-
-    def run_test(self,df_dict):
-        df_tmp1 = df_dict["frh"].loc[df_dict["frh"]["fragment_type"]==3]
-
-        if len(df_tmp1)==0:
-            return DQMTestResult(DQMResultEnum.WARNING,f"WARNING: No Self-triggered DAPHNE data found.")
-        
-        n_emptyFrames          = len(df_tmp1.loc[df_tmp1["data_size_bytes"] == 0])
-        n_filledFrames         = len(df_tmp1.loc[df_tmp1["data_size_bytes"] != 0])
-
-        if n_emptyFrames != 0:
-            return DQMTestResult(DQMResultEnum.BAD, f'{n_emptyFrames} fragments are empty ({n_filledFrames} are fine).')
-        else:
-            return DQMTestResult(DQMResultEnum.OK,f'OK')
-
-class CheckFramesInTimeWindow_DAPHNE(DQMTest):
-
-    def __init__(self):
-        super().__init__()
-        self.name = "CheckFramesInTimeWindow_DAPHNE"
-
-    def run_test(self,df_dict):
-        ...
-
-class CheckTimestampDiffs_DAPHNE(DQMTest):
-    def __init__(self):
-        super().__init__()
-        self.name = "CheckTimestampDiffs_DAPHNE"
-
-    def run_test(self, df_dict, verbose=False):
-
-        n_bad_stream, n_bad = 0, 0
-
-        if "detd_kHD_PDS_kDAPHNEStream" not in df_dict.keys() and "detd_kHD_PDS_kDAPHNE" not in df_dict.keys():
-            return DQMTestResult(DQMResultEnum.WARNING,f'WARNING: No data for DAPHNE found.')
-        
-        if "deth_kHD_PDS_kDAPHNEStream" in df_dict.keys():
-
-            tmp_df_stream  = df_dict["deth_kHD_PDS_kDAPHNEStream"]
-            tmp_df_stream["ts_check"] = tmp_df_stream.apply(lambda x: 1 if (len(x.ts_diffs_vals)!=1) else 0, axis=1)
-            n_bad_stream = tmp_df_stream["ts_check"].sum()
-
-        if "deth_kHD_PDS_kDAPHNE" in df_dict.keys():
-
-            tmp_df  = df_dict["deth_kHD_PDS_kDAPHNEStream"]
-            tmp_df["ts_check"] = tmp_df.apply(lambda x: 1 if (len(x.ts_diffs_vals)!=1) else 0, axis=1)
-            n_bad = tmp_df["ts_check"].sum()
-
-        if n_bad == 0 and n_bad_stream == 0:
-            return DQMTestResult(DQMResultEnum.OK,f'OK')
-        else:
-            if verbose:
-                if n_bad != 0:
-                    print(tabulate(tmp_df.reset_index()[["trigger","sequence","ts_diffs_vals", "ts_diffs_counts", "ts_check"]],
-                                   headers=["record","sequence","ts_diffs","ts_diffs_counts","Check"],
-                                   showindex=False,tablefmt='pretty',floatfmt=".2f"))
-                elif n_bad_stream != 0:
-                    print(tabulate(tmp_df_stream.reset_index()[["trigger","sequence","ts_diffs_vals", "ts_diffs_counts", "ts_check"]],
-                                   headers=["record","sequence","ts_diffs","ts_diffs_counts","Check"],
-                                   showindex=False,tablefmt='pretty',floatfmt=".2f"))
-
-            return DQMTestResult(DQMResultEnum.BAD, f'{n_bad+n_bad_stream} links fail TS difference check.')
-
-
-class CheckADCData_DAPHNE(DQMTest):
-
-    def __init__(self):
-        super().__init__()
-        self.name = "CheckADCData_DAPHNE"
-    
-    def run_test(self, df_dict):
-
-        if "detd_kHD_PDS_kDAPHNEStream" not in df_dict.keys() and "detd_kHD_PDS_kDAPHNE" not in df_dict.keys():
-            return DQMTestResult(DQMResultEnum.WARNING,f'WARNING: No data for DAPHNE found.')
-        
-        elif "detd_kHD_PDS_kDAPHNEStream" in df_dict.keys():
-
-            tmp_df_stream  = df_dict["detd_kHD_PDS_kDAPHNEStream"]
-            means = np.array(df_dict["detd_kHD_PDS_kDAPHNEStream"]["adc_mean"])
-            rmss  = np.array(df_dict["detd_kHD_PDS_kDAPHNEStream"]["adc_rms"])
-
-            if np.any(means) == 0 or np.any(rmss) == 0:
-
-                n_bad_means_stream = len(means[np.where(means == 0)])
-                n_bad_rmss_stream  = len(rmss[np.where(rmss == 0)])
-                return DQMTestResult(DQMResultEnum.BAD, f'{np.max(n_bad_means_stream, n_bad_rmss_stream)} channels have problems')
-            
-            else:
-                return DQMTestResult(DQMResultEnum.OK,f'OK')
-            
-        else:
-
-            tmp_df_stream  = df_dict["detd_kHD_PDS_kDAPHNE"]
-            means = np.array(df_dict["detd_kHD_PDS_kDAPHNE"]["adc_mean"])
-            rmss  = np.array(df_dict["detd_kHD_PDS_kDAPHNE"]["adc_rms"])
-
-            if np.any(means) == 0 or np.any(rmss) == 0:
-
-                n_bad_means = len(means[np.where(means == 0)])
-                n_bad_rmss  = len(rmss[np.where(rmss == 0)])
-                return DQMTestResult(DQMResultEnum.BAD, f'{np.max(n_bad_means, n_bad_rmss)} channels have problems')
-            
-            else:
-                return DQMTestResult(DQMResultEnum.OK,f'OK')
-
-
-
